+++ lib/src/resolver/x_resolver_actions.dart
@@ -1,5 +1,4 @@
 import 'package:equatable/equatable.dart';
-import 'package:x_router/src/route/x_page_builder.dart';
 
 abstract class XResolverAction {
   const XResolverAction();
@@ -19,26 +18,4 @@ class Redirect extends XResolverAction with EquatableMixin {
   const Redirect(this.target);
   @override
   List<Object?> get props => [target];
-}
-
-/// To stop the resolving process and access the target without going to
-/// the next resolver
-class ByPass extends XResolverAction with EquatableMixin {
-  const ByPass();
-  @override
-  List<Object?> get props => [];
-}
-
-/// this is useful for not redirecting to a loading page when the state is unknow
-/// the canonical example is for authentication status
-///
-/// The alternative is to redirect to a loading page and pass it the
-/// target page as data. Then when the resolver resolves, the data is used
-/// to redirect. That process is slightly more involved and error prone
-/// than just using Loading
-class Loading extends XResolverAction with EquatableMixin {
-  final XPageBuilder loadingScreenBuilder;
-  const Loading(this.loadingScreenBuilder);
-  @override
-  List<Object?> get props => [loadingScreenBuilder];
 }